+++ packages/apps/explorer/src/components/navbar/styles.css.ts
@@ -1,5 +1,5 @@
 import { globalStyle, style } from '@vanilla-extract/css';
-import { $$pageWidth } from '../styles.css';
+import { $$pageWidth } from '../layout/styles.css';
 
 export const navbarWrapperClass = style({});
