+++ packages/apps/explorer/src/pages/block/[hash].tsx
@@ -2,6 +2,7 @@ import type { Transaction } from '@/__generated__/sdk';
 import { useBlockQuery } from '@/__generated__/sdk';
 import CompactTransactionsTable from '@/components/compact-transactions-table/compact-transactions-table';
 import DataRenderComponent from '@/components/data-render-component/data-render-component';
+import SearchLayout from '@/components/layout/search-layout/search-layout';
 import { Badge, TabItem, Tabs } from '@kadena/react-ui';
 
 import { useRouter } from 'next/router';
@@ -34,7 +35,7 @@ const Block: React.FC = () => {
   });
 
   return (
-    <>
+    <SearchLayout>
       {loading && <div>Loading...</div>}
       {error && <div>Error: {error.message}</div>}
       {data && data.block && (
@@ -145,7 +146,7 @@ const Block: React.FC = () => {
           </TabItem>
         </Tabs>
       )}
-    </>
+    </SearchLayout>
   );
 };
